Testbrugere indsat i seedingDb
Der er indsat testbrugere i seedingDb filen.

+++ app/routes/events.$eventId.destroy.jsx
@@ -3,7 +3,8 @@ import { auth } from "../sessions/auth.server";
 import { redirect } from "@remix-run/node";
 
 
-export async function loader({ request, params }) {
+export async function loader({ request }) {
+
     await auth.isAuthenticated(request, {
         failureRedirect: "/signin",
     });
@@ -15,9 +16,15 @@ export async function loader({ request, params }) {
 
 export async function action({ request, params }) {
 
-    await auth.isAuthenticated(request, {
+    const user = await auth.isAuthenticated(request, {
         failureRedirect: "/signin",
     }); 
+
+    const eventToDelete = await mongoose.models.Event.findById(params.eventId);
+
+    if (eventToDelete.createdBy.toString() !== user._id.toString()) {
+        return redirect(`/events/${params.eventId}`);
+    }
 
     await mongoose.models.Event.findByIdAndDelete(params.eventId);
